Lab13/Lab13/src/main/java/by/belstu/lab13/Builder/UserBuilder.java
package by.belstu.lab13.Builder;

import by.belstu.lab13.Models.User;
import by.belstu.lab13.Exceptions.RepositoryException;
import by.belstu.lab13.Repository.dbconstants.UserTableConstants;

import java.sql.ResultSet;
import java.sql.SQLException;

public class UserBuilder implements Builder<User> {
    @Override
    public User build(ResultSet resultSet) throws RepositoryException {
        try {
            int id = resultSet.getInt(UserTableConstants.ID.getFieldName());
            String login = resultSet.getString(UserTableConstants.LOGIN.getFieldName());
            byte[] password = resultSet.getBytes(UserTableConstants.PASSWORD.getFieldName());
            return new User(login, password);
        }
        catch (SQLException exception) {
            throw new RepositoryException(exception.getMessage(), exception);
        }
    }
}
